refactor: update the builder code struct
