+++ packages/solutions/app-tools/src/builder/builder-webpack/plugins/webpack/index.ts
@@ -1,0 +1,3 @@
+export * from './HtmlAsyncChunkPlugin';
+export * from './HtmlBottomTemplate';
+export * from './RouterPlugin';
